requirements.txt
flask
requests
python-pptx
openai
python-dotenv